Fix -Wconversion warnings: add explicit casts for size/arithmetic operations
- ethics_aware_confidence_detector.cpp: Cast size to float before multiplication
  - Line 336: Cast hedge_words.size() to float
- multi_lora_manager.cpp: Cast size to int for round-robin modulo operation
  - Line 2097: Cast num_gpus to int for round-robin index calculation
- production_validator.cpp: Cast size to double for divisions
  - Line 239: Cast latencies.size() to double for average calculation
  - Line 600: Cast latencies.size() to double for average calculation

+++ src/llm/ethics_aware_confidence_detector.cpp
@@ -333,7 +333,7 @@ float EthicsAwareConfidenceDetector::evaluateTransparency(const std::string& tex
     float score = 0.5f; // Base score
     
     // Add for hedge words (up to 0.3)
-    score += std::min(0.3f, hedge_words.size() * 0.05f);
+    score += std::min(0.3f, static_cast<float>(hedge_words.size()) * 0.05f);
     
     // Add for limitation acknowledgments (up to 0.2)
     score += std::min(0.2f, limitation_count * 0.1f);
